+++ app/src/main/java/org/simple/clinic/search/PatientSearchValidationError.kt
@@ -1,5 +1,0 @@
-package org.simple.clinic.search
-
-enum class PatientSearchValidationError(val analyticsName: String) {
-  INPUT_EMPTY("Patient Search:No input")
-}