+++ app/src/main/java/org/simple/clinic/search/results/PatientSearchResultsUi.kt
@@ -1,4 +1,0 @@
-package org.simple.clinic.search.results
-
-interface PatientSearchResultsUi {
-}