+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Bump Mobius to v1.5.3
 - Material Theming Migration
 - Show patients with national health id when searching with from scan screen
+- Remove old patient search
 
 ### Features
 - [In Progress: 06 May 2021] Add NHID support in `ScanSimpleIdScreen`